Refactor project into modular scripts
